requirements.txt
customtkinter==5.2.2
matplotlib==3.8.4
mediapipe==0.10.14
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
Pillow==10.3.0
tensorflow==2.15.0
tensorflow_intel==2.15.0
